Begin initial POC for multi-serve
